backend/requirements.txt
# FastAPI核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.12.1
pgvector==0.2.4

# 向量存储和RAG相关
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2

# 文档处理
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# AI模型集成
openai==1.3.7
httpx==0.25.2

# 工具库
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3

# 邮件发送和认证相关
fastapi-mail==1.4.1
redis==5.0.1
bcrypt==4.1.2
PyJWT==2.8.0

# 开发工具
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# 可选可观测性与监控（根据配置启用）
sentry-sdk[fastapi]==2.43.0
prometheus-fastapi-instrumentator==2.0.0
